# rewrite definition