- faster clamp
- use macro to implement default `HasBlack`/`HasWhite` traits

```
before
test affine::test::bench_rotate_bilinear                              ... bench:     624,051 ns/iter (+/- 13,916)
test filter::test::bench_horizontal_filter                            ... bench:   9,157,554 ns/iter (+/- 107,757)
test filter::test::bench_separable_filter                             ... bench:   6,541,081 ns/iter (+/- 43,448)
test filter::test::bench_vertical_filter                              ... bench:   9,359,800 ns/iter (+/- 110,387)

after
test affine::test::bench_rotate_bilinear                              ... bench:     524,959 ns/iter (+/- 2,893)
test filter::test::bench_horizontal_filter                            ... bench:   8,201,409 ns/iter (+/- 221,800)
test filter::test::bench_separable_filter                             ... bench:   5,883,217 ns/iter (+/- 45,865)
test filter::test::bench_vertical_filter                              ... bench:   8,478,728 ns/iter (+/- 50,299)
```