@theotherphil 2016-01-07
Nice speedups again! Are the "as $from" bits necessary after you've already cast in the macro call?